[BUGFIX] Don't create two visitor-records if a new lead enters a page where is also a contextual content plugin on it
This happened because the Pi2 ControllerAction also used the VisitorFactory. And this creates a new record - at the same time when the PageRequest also calls the VisitorFactory.
Now we just search in Pi2 for a visitor in database with the same cookieId. If there is no visitor yet (on the very first page call), null is passed to ContextualContentService and this should return the default content element.